+++ diesel/src/types/impls/date_and_time/chrono.rs
@@ -107,17 +107,17 @@ mod tests {
     use self::dotenv::dotenv;
 
     use ::select;
-    use connection::Connection;
+    use connection::{Connection, PgConnection};
     use expression::dsl::{sql, now};
     use prelude::*;
     use types::{Date, Time, Timestamp};
 
-    fn connection() -> Connection {
+    fn connection() -> PgConnection {
         dotenv().ok();
 
         let connection_url = ::std::env::var("DATABASE_URL").ok()
             .expect("DATABASE_URL must be set in order to run tests");
-        Connection::establish(&connection_url).unwrap()
+        PgConnection::establish(&connection_url).unwrap()
     }
 
     #[test]
