+++ diesel_tests/build.rs
@@ -28,13 +28,14 @@ mod inner {
 extern crate diesel;
 extern crate dotenv;
 use diesel::*;
+use diesel::connection::PgConnection;
 use dotenv::dotenv;
 
 fn main() {
     dotenv().ok();
     let database_url = ::std::env::var("DATABASE_URL")
         .expect("DATABASE_URL must be set to run tests");
-    let connection = Connection::establish(&database_url).unwrap();
+    let connection = PgConnection::establish(&database_url).unwrap();
     migrations::run_pending_migrations(&connection).unwrap();
     inner::main();
 }
