+++ diesel_tests/tests/compile-fail/cannot_mix_aggregate_and_non_aggregate_selects.rs
@@ -3,6 +3,7 @@ extern crate diesel;
 
 use diesel::*;
 use diesel::expression::count;
+use diesel::connection::PgConnection;
 
 table! {
     users {
@@ -13,7 +14,7 @@ table! {
 fn main() {
     use self::users::dsl::*;
 
-    let connection = Connection::establish("").unwrap();
+    let connection = PgConnection::establish("").unwrap();
     let source = users.select((id, count(users.star())));
     //~^ ERROR E0277
     //~| ERROR E0277
